Move naturalSort and naturalSortByProperty to public.
Also had to move similaritySortFactory to its own file to prevent circular import issues.
