+++ packages/components/src/internal/components/gridbar/ViewSelector.tsx
@@ -19,7 +19,8 @@ import { List } from 'immutable';
 
 import { gridSelectView } from '../../actions';
 import { QueryGridModel, ViewInfo } from '../base/models/model';
-import { generateId, naturalSort } from '../../util/utils';
+import { generateId } from '../../util/utils';
+import { naturalSort } from '../../..';
 
 const emptyList = List<ReactNode>();
