+++ packages/components/src/internal/components/navigation/ProductMenuSection.tsx
@@ -17,10 +17,10 @@ import React, { Component, ReactNode } from 'react';
 import { Record } from 'immutable';
 
 import { AppURL } from '../../url/AppURL';
-import { naturalSort } from '../../util/utils';
 
 import { MenuSectionModel } from './model';
 import { createProductUrlFromParts, getHref } from './utils';
+import { naturalSort } from '../../..';
 
 export class MenuSectionConfig extends Record({
     emptyText: undefined,
